Reorganize the package a bit
